+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- [#4](https://github.com/pgerke/ngx-mat-table-multi-sort/issues/4): Support Angular 18
 - [#20](https://github.com/pgerke/ngx-mat-table-multi-sort/issues/20): Expanded the documentation
 
 ## [19.0.0](https://github.com/pgerke/ngx-mat-table-multi-sort/releases/tag/19.0.0) - 29.12.2024